Challenge accepted!
I made two examples.

The first one is if the user needs to keep the response as it is, for example, to fit a already live app/service (Probably not the case). This could mean that the service is already expecting code 200 every time and, what distinguishes is the "status" property.

The second one is way shorter then the first one, It uses "one line" if/return response, but keeping it completly readable. 'status' key can be added to the arrays if needed. Omitted to be more readable for a junior, but can be easily added.

Ideally, request could be a custom request, with validation on another file, but I wanted to stick to one file, to keep the simplicity of the task.

Waiting for your comments/opinion!